+++ src/main/java/gtclassic/material/GTMaterial.java
@@ -54,7 +54,6 @@ public class GTMaterial {
 	public static final GTMaterial Pyrite = new GTMaterial("Pyrite", 150, 120, 40, dust);
 	public static final GTMaterial Ruby = new GTMaterial("Ruby", 255, 75, 75, rubyAll);
 	public static final GTMaterial Sapphire = new GTMaterial("Sapphire", 75, 75, 200, sapphireAll);
-	public static final GTMaterial SapphireGreen = new GTMaterial("GreenSapphire", 100, 200, 130, dust);
 	public static final GTMaterial Silicon = new GTMaterial("Silicon", 60, 60, 80, dust, ingot);
 	public static final GTMaterial Sodalite = new GTMaterial("Sodalite", 20, 20, 255, dust);
 	public static final GTMaterial Sodium = new GTMaterial("Sodium", 0, 38, 255, fluid);