+++ CHANGELOG.md
@@ -15,7 +15,11 @@
 - Fixed a naming issue with the super conductor item, might have to recraft some if you had any lying around.
 - Fixed adv circuts recipes eating tons of metal with no added benefit, now you can get 2x adv circuits.
 - Added config option to show all possible fluids as GTC tubes instead of just GTC based fluids.
+- Fixed GTC ores not responding to fortune.
+- Changed mining laser recipe to be somewhat like GT2, as it is very OP these days.
+- Changed recipe for IC2C Battery Station, as it is also very OP.
 - Added GT1 recipe for Ic2c nuke if the nuke is enabled in ic2c configs.
+- Added original GT recipe for Buildcraft Quarry - can be disabled in configs.
 
 # 1.04
 - WARNING! In the next version (1.0.5) the Blast Furnace will become part of GTC Expansion (by Trinsdar)